requirements.txt
appdirs==1.4.3
cffi==1.12.3
Django==2.2.3
django-crispy-forms==1.4.0
django-registration-redux==1.1
djangorestframework==3.4.0
docx2txt==0.8
future==0.17.1
gunicorn==19.7.0
lxml==4.3.4
mime==0.1.0
numpy==1.16.4
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.8.3
pycparser==2.19
pylokit==0.8.1
pyparsing==2.2.0
PyPDF2==1.26.0
pytesseract==0.2.7
python-docx==0.8.10
pytz==2019.1
scipy==1.3.0
six==1.10.0
sqlparse==0.3.0
templated-docs==0.3.1
